Refactor 'nmapSessionObj' entity into its own file
